docker/bot/requirements.txt
# Основной Telegram бот
python-telegram-bot>=20.0

# Обработка изображений и распознавание текста
pillow>=9.5.0
pytesseract>=0.3.10

# Работа с документами
python-docx>=0.8.11

# Интеграция с Git
gitpython>=3.1.30

# Переменные окружения
python-dotenv>=1.0.0

# Логирование и обработка ошибок
loguru>=0.7.0

# Веб-запросы (опционально для интеграций)
requests>=2.31.0

# Работа с файловой системой
pathlib>=1.0.1

# Обработка дат и времени
python-dateutil>=2.8.2

# Операционная система и процессы
psutil>=5.9.5

# Работа с SSH
paramiko>=3.3.1

# Асинхронные операции
aiohttp>=3.9.1
asyncssh>=2.14.0

# Работа с базами данных SQLite
aiosqlite>=0.19.0

# Планировщик задач
apscheduler>=3.10.4